Fix And Move file
