+++ domain/service/curriculum_service.go
@@ -7,7 +7,7 @@ import (
 
 	"tat_gogogo/domain/model"
 	"tat_gogogo/domain/repository"
-	"tat_gogogo/glob/logs"
+	"tat_gogogo/utilities/logs"
 	"tat_gogogo/utilities/decoder"
 	"tat_gogogo/utilities/httcli"
